Refactor HAClusterValidator classes for improved structure and functionality
