+++ src/modules/get_pcmk_properties_scs.py
@@ -15,13 +15,11 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.facts.compat import ansible_facts
 
 try:
-    from ansible.module_utils.sap_automation_qa import SapAutomationQA
-    from ansible.module_utils.enums import OperatingSystemFamily, Parameters, TestStatus
-    from ansible.module_utils.commands import CIB_ADMIN
+    from ansible.module_utils.get_pcmk_properties import BaseHAClusterValidator
+    from ansible.module_utils.enums import OperatingSystemFamily, TestStatus
 except ImportError:
-    from src.module_utils.sap_automation_qa import SapAutomationQA
-    from src.module_utils.enums import OperatingSystemFamily, Parameters, TestStatus
-    from src.module_utils.commands import CIB_ADMIN
+    from src.module_utils.get_pcmk_properties import BaseHAClusterValidator
+    from src.module_utils.enums import OperatingSystemFamily, TestStatus
 
 
 DOCUMENTATION = r"""
@@ -152,26 +150,13 @@ details:
 """
 
 
-class HAClusterValidator(SapAutomationQA):
+class HAClusterValidator(BaseHAClusterValidator):
     """
-    Validates High Availability cluster configurations.
+    Validates High Availability cluster configurations for SAP ASCS/ERS.
 
-    This class validates Pacemaker cluster configurations against predefined
-    standards for SAP HANA deployments. It checks both basic cluster properties
-    and resource-specific configurations.
-
-    Attributes:
-        BASIC_CATEGORIES (Dict): Mapping of basic configuration categories to their XPaths
-        RESOURCE_CATEGORIES (Dict): Mapping of resource types to their XPaths
+    This class extends BaseHAClusterValidator to provide ASCS/ERS-specific validation
+    functionality including NFS provider handling and ASCS/ERS resource configurations.
     """
-
-    BASIC_CATEGORIES = {
-        "crm_config": (".//cluster_property_set", "CRM_CONFIG_DEFAULTS"),
-        "rsc_defaults": (".//meta_attributes", "RSC_DEFAULTS"),
-        "op_defaults": (".//meta_attributes", "OP_DEFAULTS"),
-    }
-
-    CONSTRAINTS_CATEGORIES = (".//*", "CONSTRAINTS_DEFAULTS")
 
     RESOURCE_CATEGORIES = {
         "sbd_stonith": ".//primitive[@type='external/sbd']",
@@ -190,380 +175,111 @@ class HAClusterValidator(SapAutomationQA):
         virtual_machine_name: str,
         constants: dict,
         fencing_mechanism: str,
+        cib_output: str,
         nfs_provider=None,
         category=None,
     ):
-        super().__init__()
-        self.os_type = os_type.value.upper()
-        self.category = category
-        self.sid = sid
+        super().__init__(
+            os_type=os_type,
+            sid=sid,
+            virtual_machine_name=virtual_machine_name,
+            constants=constants,
+            fencing_mechanism=fencing_mechanism,
+            category=category,
+            cib_output=cib_output,
+        )
         self.scs_instance_number = scs_instance_number
         self.ers_instance_number = ers_instance_number
-        self.virtual_machine_name = virtual_machine_name
-        self.fencing_mechanism = fencing_mechanism
-        self.constants = constants
         self.nfs_provider = nfs_provider
         self.parse_ha_cluster_config()
 
-    def _get_expected_value(self, category, name):
+    def _get_expected_value_for_category(self, category, subcategory, name, op_name):
         """
-        Get expected value for basic configuration parameters.
+        Get expected value based on category type with SCS-specific logic.
 
-        :param category: Category of the parameter
+        :param category: The category of the configuration parameter.
         :type category: str
-        :param name: Name of the parameter
-        :type name: str
-        :return: Expected value of the parameter
-        :rtype: str
-        """
-        _, defaults_key = self.BASIC_CATEGORIES[category]
-
-        fence_config = self.constants["VALID_CONFIGS"].get(self.fencing_mechanism, {})
-        os_config = self.constants["VALID_CONFIGS"].get(self.os_type, {})
-
-        return fence_config.get(name) or os_config.get(name, self.constants[defaults_key].get(name))
-
-    def _get_resource_expected_value(self, resource_type, section, param_name, op_name=None):
-        """
-        Get expected value for resource-specific configuration parameters.
-
-        :param resource_type: Type of the resource (e.g., stonith, ipaddr)
-        :type resource_type: str
-        :param section: Section of the resource (e.g., meta_attributes, operations)
-        :type section: str
-        :param param_name: Name of the parameter
-        :type param_name: str
-        :param op_name: Name of the operation (if applicable)
-        :type op_name: str
-        :return: Expected value of the parameter
-        :rtype: str
-        """
-        resource_defaults = (
-            self.constants["RESOURCE_DEFAULTS"].get(self.os_type, {}).get(resource_type, {})
-        )
-
-        if section == "meta_attributes":
-            return resource_defaults.get("meta_attributes", {}).get(param_name)
-        elif section == "operations":
-            ops = resource_defaults.get("operations", {}).get(op_name, {})
-            return ops.get(param_name)
-        elif section == "instance_attributes":
-            return resource_defaults.get("instance_attributes", {}).get(param_name)
-        return None
-
-    def _create_parameter(
-        self,
-        category,
-        name,
-        value,
-        expected_value=None,
-        id=None,
-        subcategory=None,
-        op_name=None,
-    ):
-        """
-        Create a Parameters object for a given configuration parameter.
-
-        :param category: Category of the parameter
-        :type category: str
-        :param name: Name of the parameter
-        :type name: str
-        :param value: Value of the parameter
-        :type value: str
-        :param expected_value: Expected value of the parameter
-        :type expected_value: str
-        :param id: ID of the parameter (optional)
-        :type id: str
-        :param subcategory: Subcategory of the parameter (optional)
+        :param subcategory: The subcategory of the configuration parameter.
         :type subcategory: str
-        :param op_name: Operation name (optional)
+        :param name: The name of the configuration parameter.
+        :type name: str
+        :param op_name: The name of the operation (if applicable).
         :type op_name: str
-        :return: Parameters object
-        :rtype: Parameters
+        :return: The expected value for the configuration parameter.
+        :rtype: str or list or dict
         """
-        status = None
-        if expected_value is None:
-            if category in self.RESOURCE_CATEGORIES or category in ["ascs", "ers"]:
-                expected_value = self._get_resource_expected_value(
-                    resource_type=category,
-                    section=subcategory,
-                    param_name=name,
-                    op_name=op_name,
-                )
-            else:
-                expected_value = self._get_expected_value(category, name)
+        if category in self.RESOURCE_CATEGORIES or category in ["ascs", "ers"]:
+            return self._get_resource_expected_value(
+                resource_type=category,
+                section=subcategory,
+                param_name=name,
+                op_name=op_name,
+            )
+        else:
+            return self._get_expected_value(category, name)
 
+    def _determine_parameter_status(self, value, expected_value):
+        """
+        Determine the status of a parameter with SCS-specific logic for NFS provider.
+
+        :param value: The actual value of the parameter.
+        :type value: str
+        :param expected_value: The expected value of the parameter.
+        :type expected_value: str or list or dict
+        :return: The status of the parameter.
+        :rtype: str
+        """
         if expected_value is None or value == "":
-            status = TestStatus.INFO.value
+            return TestStatus.INFO.value
         elif isinstance(expected_value, (str, list)):
             if isinstance(expected_value, list):
-                status = (
+                return (
                     TestStatus.SUCCESS.value
                     if str(value) in expected_value
                     else TestStatus.ERROR.value
                 )
-                expected_value = expected_value[0]
             else:
-                status = (
+                return (
                     TestStatus.SUCCESS.value
                     if str(value) == str(expected_value)
                     else TestStatus.ERROR.value
                 )
         elif isinstance(expected_value, dict):
-            expected_value = expected_value.get(self.nfs_provider, "AFS")
-            status = (
-                TestStatus.SUCCESS.value if str(value) in expected_value else TestStatus.ERROR.value
+            provider_values = expected_value.get(self.nfs_provider, expected_value.get("AFS", []))
+            return (
+                TestStatus.SUCCESS.value
+                if str(value) in provider_values
+                else TestStatus.ERROR.value
             )
-            expected_value = expected_value[0]
         else:
-            status = TestStatus.ERROR.value
+            return TestStatus.ERROR.value
 
-        return Parameters(
-            category=f"{category}_{subcategory}" if subcategory else category,
-            id=id if id else "",
-            name=name if not op_name else f"{op_name}_{name}",
-            value=value,
-            expected_value=expected_value if expected_value is not None else "",
-            status=status if status else TestStatus.ERROR.value,
-        ).to_dict()
-
-    def _parse_nvpair_elements(self, elements, category, subcategory=None, op_name=None):
+    def _parse_resources_section(self, root):
         """
-        Parse nvpair elements and return a list of Parameters objects.
+        Parse resources section with ASCS/ERS-specific logic.
 
-        :param elements: List of XML elements to parse
-        :type elements: List[ElementTree.Element]
-        :param category: Category of the parameters
-        :type category: str
-        :param subcategory: Subcategory of the parameters
-        :type subcategory: str
-        :param op_name: Operation name (if applicable)
-        :type op_name: str
-        :return: List of Parameters objects
-        :rtype: List[Parameters]
-        """
-        parameters = []
-        for nvpair in elements:
-            name = nvpair.get("name", "")
-            if name in ["passwd", "password", "login"]:
-                continue
-            else:
-                parameters.append(
-                    self._create_parameter(
-                        category=category,
-                        subcategory=subcategory,
-                        op_name=op_name,
-                        id=nvpair.get("id", ""),
-                        name=name,
-                        value=nvpair.get("value", ""),
-                    )
-                )
-        return parameters
-
-    def _parse_resource(self, element, category):
-        """
-        Parse resource-specific configuration parameters
-
-        :param element: XML element to parse
-        :type element: ElementTree.Element
-        :param category: Resource category (e.g., stonith, ipaddr)
-        :type category: str
-        :return: List of Parameters objects for the resource
-        :rtype: List[Parameters]
+        :param root: The XML root element to parse.
+        :type root: xml.etree.ElementTree.Element
+        :return: A list of parameter dictionaries.
+        :rtype: list
         """
         parameters = []
 
-        for attr in ["meta_attributes", "instance_attributes"]:
-            attr_elements = element.find(f".//{attr}")
-            if attr_elements is not None:
-                parameters.extend(
-                    self._parse_nvpair_elements(
-                        elements=attr_elements.findall(".//nvpair"),
-                        category=category,
-                        subcategory=attr,
-                    )
-                )
+        for sub_category, xpath in self.RESOURCE_CATEGORIES.items():
+            elements = root.findall(xpath)
+            for element in elements:
+                parameters.extend(self._parse_resource(element, sub_category))
 
-        operations = element.find(".//operations")
-        if operations is not None:
-            for operation in operations.findall(".//op"):
-                for op_type in ["timeout", "interval"]:
-                    parameters.append(
-                        self._create_parameter(
-                            category=category,
-                            subcategory="operations",
-                            id=operation.get("id", ""),
-                            name=op_type,
-                            op_name=operation.get("name", ""),
-                            value=operation.get(op_type, ""),
-                        )
-                    )
-        return parameters
-
-    def _parse_basic_config(self, element, category, subcategory=None):
-        """
-        Parse basic configuration parameters
-
-        :param element: XML element to parse
-        :type element: ElementTree.Element
-        :param category: Category of the parameters
-        :type category: str
-        :param subcategory: Subcategory of the parameters
-        :type subcategory: str
-        :return: List of Parameters objects for basic configuration
-        :rtype: List[Parameters]
-        """
-        parameters = []
-        for nvpair in element.findall(".//nvpair"):
-            parameters.append(
-                self._create_parameter(
-                    category=category,
-                    subcategory=subcategory,
-                    name=nvpair.get("name", ""),
-                    value=nvpair.get("value", ""),
-                    id=nvpair.get("id", ""),
-                )
-            )
-        return parameters
-
-    def _parse_os_parameters(self):
-        """
-        Parse OS-specific parameters
-
-        :return: List of Parameters objects for OS parameters
-        :rtype: List[Parameters]
-        """
-        parameters = []
-
-        os_parameters = self.constants["OS_PARAMETERS"].get("DEFAULTS", {})
-
-        for section, params in os_parameters.items():
-            for param_name, expected_value in params.items():
-                value = (
-                    self.execute_command_subprocess(command=[section, param_name])
-                    .strip()
-                    .split("\n")[0]
-                )
-                parameters.append(
-                    self._create_parameter(
-                        category="os",
-                        id=section,
-                        name=param_name,
-                        value=value,
-                        expected_value=expected_value,
-                    )
-                )
+        for group in root.findall(".//group"):
+            group_id = group.get("id", "")
+            if "ASCS" in group_id:
+                for element in group.findall(".//primitive[@type='SAPInstance']"):
+                    parameters.extend(self._parse_resource(element, "ascs"))
+            elif "ERS" in group_id:
+                for element in group.findall(".//primitive[@type='SAPInstance']"):
+                    parameters.extend(self._parse_resource(element, "ers"))
 
         return parameters
-
-    def _parse_constraints(self, root):
-        """
-        Parse constraints configuration parameters
-
-        :param root: XML root element
-        :type root: ElementTree.Element
-        :return: List of Parameters objects for constraints
-        :rtype: List[Parameters]
-        """
-        parameters = []
-        for element in root:
-            tag = element.tag
-            if tag in self.constants["CONSTRAINTS"]:
-                for attr, expected in self.constants["CONSTRAINTS"][tag].items():
-                    if element.get(attr) is not None:
-                        parameters.append(
-                            self._create_parameter(
-                                category="constraints",
-                                subcategory=tag,
-                                id=element.get("id", ""),
-                                name=attr,
-                                value=element.get(attr),
-                                expected_value=expected,
-                            )
-                        )
-                    else:
-                        continue
-            else:
-                continue
-        return parameters
-
-    def parse_ha_cluster_config(self):
-        """
-        Parse HA cluster configuration XML and return a list of properties.
-        """
-        parameters = []
-
-        for scope in [
-            "rsc_defaults",
-            "crm_config",
-            "op_defaults",
-            "constraints",
-            "resources",
-        ]:
-            self.category = scope
-            root = self.parse_xml_output(self.execute_command_subprocess(CIB_ADMIN(scope=scope)))
-            if not root:
-                continue
-
-            if self.category in self.BASIC_CATEGORIES:
-                try:
-                    xpath = self.BASIC_CATEGORIES[self.category][0]
-                    for element in root.findall(xpath):
-                        parameters.extend(self._parse_basic_config(element, self.category))
-                except Exception as ex:
-                    self.result[
-                        "message"
-                    ] += f"Failed to get {self.category} configuration: {str(ex)}"
-                    continue
-
-            elif self.category == "resources":
-                try:
-                    for sub_category, xpath in self.RESOURCE_CATEGORIES.items():
-                        elements = root.findall(xpath)
-                        for element in elements:
-                            parameters.extend(self._parse_resource(element, sub_category))
-
-                    for group in root.findall(".//group"):
-                        group_id = group.get("id", "")
-                        if "ASCS" in group_id:
-                            for element in group.findall(".//primitive[@type='SAPInstance']"):
-                                parameters.extend(self._parse_resource(element, "ascs"))
-                        elif "ERS" in group_id:
-                            for element in group.findall(".//primitive[@type='SAPInstance']"):
-                                parameters.extend(self._parse_resource(element, "ers"))
-
-                except Exception as ex:
-                    self.result[
-                        "message"
-                    ] += f"Failed to get resources configuration for {self.category}: {str(ex)}"
-                    continue
-
-            elif self.category == "constraints":
-                try:
-                    parameters.extend(self._parse_constraints(root))
-                except Exception as e:
-                    self.result["message"] += f"Failed to get constraints configuration: {str(e)}"
-                    continue
-
-        try:
-            parameters.extend(self._parse_os_parameters())
-        except Exception as ex:
-            self.result["message"] += f"Failed to get OS parameters: {str(ex)} \n"
-
-        failed_parameters = [
-            param
-            for param in parameters
-            if param.get("status", TestStatus.ERROR.value) == TestStatus.ERROR.value
-        ]
-        self.result.update(
-            {
-                "details": {"parameters": parameters},
-                "status": (
-                    TestStatus.ERROR.value if failed_parameters else TestStatus.SUCCESS.value
-                ),
-            }
-        )
-        self.result["message"] += "HA Parameter Validation completed successfully."
 
 
 def main() -> None:
@@ -579,6 +295,7 @@ def main() -> None:
             pcmk_constants=dict(type="dict"),
             fencing_mechanism=dict(type="str"),
             nfs_provider=dict(type="str", default=""),
+            cib_output=dict(type="str", required=False, default=""),
             filter=dict(type="str", required=False, default="os_family"),
         )
     )
@@ -594,6 +311,7 @@ def main() -> None:
         constants=module.params["pcmk_constants"],
         fencing_mechanism=module.params["fencing_mechanism"],
         nfs_provider=module.params.get("nfs_provider"),
+        cib_output=module.params.get("cib_output"),
     )
     module.exit_json(**validator.get_result())
