+++ src/playbook_00_ha_db_offline_tests.yml
@@ -1,0 +1,60 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+# /*---------------------------------------------------------------------------8
+# |                                                                            |
+# |                           HA Testing Framework                            |
+# |                                                                            |
+# +------------------------------------4--------------------------------------*/
+- hosts:                               localhost
+  gather_facts:                        true
+  name:                                "Setup deployer for HA Testing Framework"
+  vars_files:                          "./vars/input-api.yaml"
+  tasks:
+    - name:                            "Install python azure pacakges required"
+      become:                          true
+      ansible.builtin.pip:
+        name:
+          - ansible-runner
+          - azure-kusto-data
+          - azure-kusto-ingest
+          - azure-identity
+          - azure-mgmt-network
+          - pandas
+
+    - name:                            "Set the test group name based on the inputs"
+      ansible.builtin.set_fact:
+        test_group_name:               "{{ sap_functional_test_type_map
+                                        | selectattr('name', 'equalto', sap_functional_test_type)
+                                        | map(attribute='value') | first }}_{{ platform | upper }}"
+      run_once:                        true
+
+    - name:                            Set the test group facts
+      ansible.builtin.set_fact:
+        test_group_start_time:         "{{ now(utc=true,fmt='%Y-%m-%d %H:%M:%S') }}"
+        test_group_invocation_id:      "{{ lookup('pipe', 'uuidgen') }}"
+        test_cases:                    "{{ test_groups
+                                            | selectattr('name', 'equalto', test_group_name)
+                                            | map(attribute='test_cases') | list | flatten(levels=1)
+                                            | selectattr('enabled', 'equalto', true) | list }}"
+      run_once:                        true
+
+    - name:                            "Run test cases by including them as roles"
+      ansible.builtin.include_tasks:   "./roles/{{ test_group_name | lower }}/tasks/{{ item.task_name }}.yml"
+      loop:                            "{{ test_cases | list }}"
+      vars:
+        group_invocation_id:           "{{ test_group_invocation_id }}"
+        group_start_time:              "{{ test_group_start_time }}"
+        group_name:                    "{{ sap_functional_test_type }}"
+      when:
+                                        - test_group_name is defined
+                                        - item.enabled | default(true)
+
+    - name:                            "Render HTML report for the test group"
+      ansible.builtin.include_tasks:   "./roles/misc/tasks/render-html-report.yml"
+      when:                            test_group_name is defined
+
+    - name:                            "Debug the group_invocation_id"
+      ansible.builtin.debug:
+        msg:                           "Group invocation ID: {{ test_group_invocation_id }}"
+      when:                            test_group_invocation_id is defined
